event_organizers_service/requirements.txt
Django==3.2.10
djangorestframework==3.12.4
psycopg2-binary==2.9.3
gunicorn==20.1.0
python-consul==1.1.0
drf-spectacular==0.27.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==3.11.0